updated sample app
