+++ sample-apps/payment-customizations/extensions/payment-customization-js/locales/en.default.json
@@ -1,0 +1,4 @@
+{
+  "name": "payment-customization-js",
+  "description": "payment-customization-js"
+}
