+++ sample-apps/payment-customizations/extensions/payment-customization-js/src/run.graphql
@@ -1,4 +1,4 @@
-query Input {
+query RunInput {
   cart {
     cost {
       totalAmount {
