+++ sample-apps/payment-customizations/extensions/payment-customization-rust/.gitignore
@@ -1,3 +1,3 @@
 /target
 Cargo.lock
-.output.graphql
+*.output.graphql
