+++ sample-apps/payment-customizations/extensions/payment-customization-rust/Cargo.toml
@@ -7,7 +7,7 @@ rust-version = "1.62"
 [dependencies]
 serde = { version = "1.0.13", features = ["derive"] }
 serde_json = "1.0"
-shopify_function = "0.3.0"
+shopify_function = "0.4.0"
 graphql_client = "0.13.0"
 
 [profile.release]
