+++ sample-apps/payment-customizations/extensions/payment-customization-rust/locales/en.default.json
@@ -1,0 +1,4 @@
+{
+  "name": "payment-customization-rust",
+  "description": "payment-customization-rust"
+}
