+++ sample-apps/payment-customizations/extensions/payment-customization-rust/shopify.extension.toml
@@ -1,11 +1,21 @@
-api_version = "2023-07"
-name = "payment-customization-rust"
-type = "payment_customization"
+api_version = "2023-10"
 
-[build]
-command = "cargo wasi build --release"
-path = "target/wasm32-wasi/release/payment-customization.wasm"
+[[extensions]]
+handle = "payment-customization-rust"
+name = "t:name"
+description = "t:description"
+type = "function"
 
-[ui.paths]
-create = "/app/payment-customization/:functionId/new"
-details = "/app/payment-customization/:functionId/:id"
+  [[extensions.targeting]]
+  target = "purchase.payment-customization.run"
+  input_query = "src/run.graphql"
+  export = "run"
+
+  [extensions.build]
+  command = "cargo wasi build --release"
+  path = "target/wasm32-wasi/release/payment-customization.wasm"
+  watch = [ "src/**/*.rs" ]
+
+  [extensions.ui.paths]
+  create = "/app/payment-customization/:functionId/new"
+  details = "/app/payment-customization/:functionId/:id"
